requirements.txt
numpy==1.23.5
pandas==1.5.1
PyQt5==5.15.7
pyqtgraph==0.13.1
pyserial==3.3